+++ lib/ext/Explore.dart
@@ -7,9 +7,8 @@ import 'package:illinois/model/Laundry.dart';
 import 'package:illinois/model/MTD.dart';
 import 'package:illinois/model/StudentCourse.dart';
 import 'package:illinois/model/wellness/Appointment.dart';
-import 'package:illinois/service/DeepLink.dart';
+import 'package:illinois/model/wellness/WellnessBuilding.dart';
 import 'package:illinois/service/StudentCourses.dart';
-import 'package:illinois/service/Wellness.dart';
 import 'package:illinois/ui/academics/StudentCourses.dart';
 import 'package:illinois/ui/athletics/AthleticsGameDetailPanel.dart';
 import 'package:illinois/ui/events/CompositeEventsDetailPanel.dart';
@@ -17,6 +16,7 @@ import 'package:illinois/ui/explore/ExploreBuildingDetailPanel.dart';
 import 'package:illinois/ui/explore/ExploreDetailPanel.dart';
 import 'package:illinois/ui/explore/ExploreDiningDetailPanel.dart';
 import 'package:illinois/ui/explore/ExploreEventDetailPanel.dart';
+import 'package:illinois/ui/guide/GuideDetailPanel.dart';
 import 'package:illinois/ui/laundry/LaundryRoomDetailPanel.dart';
 import 'package:illinois/ui/mtd/MTDStopDeparturesPanel.dart';
 import 'package:illinois/ui/wellness/appointments/AppointmentDetailPanel.dart';
@@ -40,7 +40,6 @@ import 'package:geolocator/geolocator.dart' as Core;
 import 'dart:math' as math;
 
 import 'package:sprintf/sprintf.dart';
-import 'package:url_launcher/url_launcher.dart';
 
 extension ExploreExt on Explore {
 
@@ -138,6 +137,9 @@ extension ExploreExt on Explore {
     else if (exploresType == "building") {
       return Localization().getStringEx('panel.explore.item.buildings.name', 'Buildings');
     }
+    else if (exploresType == "wellnessbuilding") {
+      return Localization().getStringEx('panel.explore.item.wellnessbuildings.name', 'Resources');
+    }
     else if (exploresType == "mtdstop") {
       return Localization().getStringEx('panel.explore.item.mtd_stops.name', 'Bus Stops');
     }
@@ -234,6 +236,7 @@ extension ExploreExt on Explore {
       return (this as ExplorePOI).uiColor;
     }
     //else if (this is Building) {}
+    //else if (this is WellnessBuilding) {}
     //else if (this is Appointment) {}
     else {
       return Styles().colors?.accentColor2;
@@ -244,7 +247,6 @@ extension ExploreExt on Explore {
 
   void exploreLaunchDetail(BuildContext context, { Core.Position? initialLocationData }) {
     Route? route;
-    String? url;
     if (this is Event) {
       if ((this as Event).isGameEvent) {
         route = CupertinoPageRoute(builder: (context) => AthleticsGameDetailPanel(gameId: (this as Event).speaker, sportName: (this as Event).registrationLabel,),);
@@ -266,10 +268,10 @@ extension ExploreExt on Explore {
       route = CupertinoPageRoute(builder: (context) => AthleticsGameDetailPanel(game: this as Game),);
     }
     else if (this is Building) {
-      url = Wellness().mentalHealthBuildingUrl(buildingId: (this as Building).id);
-      if (url == null) {
-        route = CupertinoPageRoute(builder: (context) => ExploreBuildingDetailPanel(building: this as Building),);
-      }
+      route = CupertinoPageRoute(builder: (context) => ExploreBuildingDetailPanel(building: this as Building),);
+    }
+    else if (this is WellnessBuilding) {
+      route = CupertinoPageRoute(builder: (context) => GuideDetailPanel(guideEntryId: (this as WellnessBuilding).guideId),);
     }
     else if (this is MTDStop) {
       route = CupertinoPageRoute(builder: (context) => MTDStopDeparturesPanel(stop: this as MTDStop,),);
@@ -287,17 +289,6 @@ extension ExploreExt on Explore {
       route = CupertinoPageRoute(builder: (context) => ExploreDetailPanel(explore: this, initialLocationData: initialLocationData,),);
     }
 
-    if (url != null) {
-      if (DeepLink().isAppUrl(url)) {
-        DeepLink().launchUrl(url);
-      }
-      else {
-        Uri? uri = Uri.tryParse(url);
-        if (uri != null) {
-          launchUrl(uri, mode: LaunchMode.externalApplication);
-        }
-      }
-    }
     if (route != null) {
       Navigator.push(context, route);
     }
@@ -305,35 +296,6 @@ extension ExploreExt on Explore {
 }
 
 extension ExploreMap on Explore {
-
-  String? get mapMarkerAssetName {
-    if (this is Event) {
-      return 'images/map-marker-group-event.png';
-    }
-    else if (this is Dining) {
-      return 'images/map-marker-group-dining.png';
-    }
-    else if (this is LaundryRoom) {
-      return 'images/map-marker-group-laundry.png';
-    }
-    else if (this is Game) {
-      return 'images/map-marker-group-game.png';
-    }
-    else if (this is MTDStop) {
-      return 'images/map-marker-group-mtd-stop.png';
-    }
-    else if (this is StudentCourse) {
-      return 'images/map-marker-group-event.png';
-    }
-    else if (this is ExplorePOI) {
-      return 'images/map-marker-group-poi.png';
-    }
-    //else if (this is Building) {}
-    //else if (this is Appointment) {}
-    else {
-      return 'images/map-marker-group-laundry.png';
-    }
-  }
 
   Color? get mapMarkerColor => uiColor ?? unknownMarkerColor;
   static Color? get unknownMarkerColor => Styles().colors?.accentColor2;
@@ -371,10 +333,13 @@ extension ExploreMap on Explore {
       return exploreLocationDescription;
     }
     else if (this is Building) {
-      return (this as Building).address1;
+      return exploreSubTitle;
+    }
+    else if (this is WellnessBuilding) {
+      return exploreSubTitle;
     }
     else if (this is Appointment) {
-      return (this as Appointment).location?.title;
+      return exploreSubTitle;
     }
     else {
       return null;
@@ -406,6 +371,9 @@ extension ExploreMap on Explore {
     else if (this is Building) {
       return sprintf('%s Buildings', [count]);
     }
+    else if (this is WellnessBuilding) {
+      return sprintf('%s Resources', [count]);
+    }
     else if (this is Appointment) {
       return sprintf('%s Appointments', [count]);
     }
@@ -426,6 +394,9 @@ extension ExploreMap on Explore {
     Building? building;
     if (this is Building) {
       building = this as Building;
+    }
+    else if (this is WellnessBuilding) {
+      building = (this as WellnessBuilding).building;
     }
     else if (this is StudentCourse) {
       building = (this as StudentCourse).section?.building;